@@ -1,18 +1,23 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <Import Project="..\..\common.props" />
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <RootNamespace>TT.SoMall</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\TT.SoMall.Domain\TT.SoMall.Domain.csproj" />
-    <PackageReference Include="Volo.Abp.Identity.Application" Version="0.16.0" />
-    <PackageReference Include="Volo.Abp.PermissionManagement.Application" Version="0.16.0" />
-    <PackageReference Include="Volo.Abp.TenantManagement.Application" Version="0.16.0" />
-    <PackageReference Include="Volo.Abp.FeatureManagement.Application" Version="0.16.0" />
+    <ProjectReference Include="..\TT.SoMall.Application.Contracts\TT.SoMall.Application.Contracts.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Volo.Abp.Account.Application" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.Identity.Application" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.PermissionManagement.Application" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.TenantManagement.Application" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.FeatureManagement.Application" Version="2.0.1" />
   </ItemGroup>
 
 </Project>
