@@ -1,25 +1,27 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <Import Project="..\..\common.props" />
-  
+
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <RootNamespace>TT.SoMall</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Volo.Abp.Identity.Domain" Version="0.16.0" />
-    <PackageReference Include="Volo.Abp.PermissionManagement.Domain.Identity" Version="0.16.0" />
-    <PackageReference Include="Volo.Abp.BackgroundJobs.Domain" Version="0.16.0" />
-    <PackageReference Include="Volo.Abp.AuditLogging.Domain" Version="0.16.0" />
-    <PackageReference Include="Volo.Abp.TenantManagement.Domain" Version="0.16.0" />
-    <PackageReference Include="Volo.Abp.FeatureManagement.Domain" Version="0.16.0" />
-    <PackageReference Include="Volo.Abp.Auditing" Version="0.16.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.DataAnnotations" Version="2.2.0" />
+    <ProjectReference Include="..\TT.SoMall.Domain.Shared\TT.SoMall.Domain.Shared.csproj" />
   </ItemGroup>
 
   <ItemGroup>
-    <EmbeddedResource Include="Localization\SoMall\*.json" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.DataAnnotations" Version="2.2.0" />
+    <PackageReference Include="Volo.Abp.Identity.Domain" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.PermissionManagement.Domain.Identity" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.IdentityServer.Domain" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.PermissionManagement.Domain.IdentityServer" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.BackgroundJobs.Domain" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.AuditLogging.Domain" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.TenantManagement.Domain" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.FeatureManagement.Domain" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.SettingManagement.Domain" Version="2.0.1" />
   </ItemGroup>
 
 </Project>
