@@ -1,17 +1,44 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <Import Project="..\..\common.props" />
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <RootNamespace>TT.SoMall</RootNamespace>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <RootNamespace>TT.SoMall.Web</RootNamespace>
     <AssetTargetFallback>$(AssetTargetFallback);portable-net45+win8+wp8+wpa81;</AssetTargetFallback>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
     <GenerateBindingRedirectsOutputType>true</GenerateBindingRedirectsOutputType>
     <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
     <PreserveCompilationContext>true</PreserveCompilationContext>
     <MvcRazorExcludeRefAssembliesFromPublish>false</MvcRazorExcludeRefAssembliesFromPublish>
+    <PreserveCompilationReferences>true</PreserveCompilationReferences>
+    <UserSecretsId>TT-SoMall-4681b4fd-151f-4221-84a4-929d86723e4c</UserSecretsId>
   </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Serilog.AspNetCore" Version="3.2.0" />
+    <PackageReference Include="Serilog.Sinks.File" Version="4.1.0" />
+    <PackageReference Include="Serilog.Sinks.Async" Version="1.4.0" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.OpenIdConnect" Version="3.1.0" />
+    <PackageReference Include="Microsoft.Extensions.Caching.StackExchangeRedis" Version="3.1.0" />
+    <PackageReference Include="Microsoft.AspNetCore.DataProtection.StackExchangeRedis" Version="3.1.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Volo.Abp.Autofac" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.AspNetCore.Mvc.Client" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.AspNetCore.Authentication.OAuth" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.AspNetCore.Mvc.UI.Theme.Basic" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.AspNetCore.Serilog" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.Http.Client.IdentityModel" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.AspNetCore.Mvc.UI.Theme.Basic" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.Identity.Web" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.TenantManagement.Web" Version="2.0.1" />
+    <PackageReference Include="Volo.Abp.FeatureManagement.Web" Version="2.0.1" />
+    <ProjectReference Include="..\TT.SoMall.HttpApi\TT.SoMall.HttpApi.csproj" />
+    <ProjectReference Include="..\TT.SoMall.HttpApi.Client\TT.SoMall.HttpApi.Client.csproj" />
+  </ItemGroup>
 
   <ItemGroup>
     <Compile Remove="Logs\**" />
