@@ -6,7 +6,9 @@ namespace TT.SoMall
     {
         public SoMallApplicationAutoMapperProfile()
         {
-            //Configure your AutoMapper mapping configuration here...
+            /* You can configure your AutoMapper mapping configuration here.
+             * Alternatively, you can split your mapping configurations
+             * into multiple profile classes for a better organization. */
         }
     }
 }
