@@ -1,48 +1,37 @@
-﻿using TT.SoMall.Localization.SoMall;
-using Volo.Abp.Auditing;
+﻿using TT.SoMall.MultiTenancy;
 using Volo.Abp.AuditLogging;
 using Volo.Abp.BackgroundJobs;
 using Volo.Abp.FeatureManagement;
 using Volo.Abp.Identity;
-using Volo.Abp.Localization;
-using Volo.Abp.Localization.Resources.AbpValidation;
+using Volo.Abp.IdentityServer;
 using Volo.Abp.Modularity;
 using Volo.Abp.MultiTenancy;
 using Volo.Abp.PermissionManagement.Identity;
+using Volo.Abp.PermissionManagement.IdentityServer;
+using Volo.Abp.SettingManagement;
 using Volo.Abp.TenantManagement;
-using Volo.Abp.VirtualFileSystem;
 
 namespace TT.SoMall
 {
     [DependsOn(
+        typeof(SoMallDomainSharedModule),
+        typeof(AbpAuditLoggingDomainModule),
+        typeof(AbpBackgroundJobsDomainModule),
+        typeof(AbpFeatureManagementDomainModule),
         typeof(AbpIdentityDomainModule),
         typeof(AbpPermissionManagementDomainIdentityModule),
-        typeof(AbpAuditingModule),
-        typeof(BackgroundJobsDomainModule),
-        typeof(AbpAuditLoggingDomainModule),
-        typeof(AbpTenantManagementDomainModule),
-        typeof(AbpFeatureManagementDomainModule)
+        typeof(AbpIdentityServerDomainModule),
+        typeof(AbpPermissionManagementDomainIdentityServerModule),
+        typeof(AbpSettingManagementDomainModule),
+        typeof(AbpTenantManagementDomainModule)
         )]
     public class SoMallDomainModule : AbpModule
     {
         public override void ConfigureServices(ServiceConfigurationContext context)
         {
-            Configure<VirtualFileSystemOptions>(options =>
+            Configure<AbpMultiTenancyOptions>(options =>
             {
-                options.FileSets.AddEmbedded<SoMallDomainModule>("TT.SoMall");
-            });
-
-            Configure<AbpLocalizationOptions>(options =>
-            {
-                options.Resources
-                    .Add<SoMallResource>("en")
-                    .AddBaseTypes(typeof(AbpValidationResource))
-                    .AddVirtualJson("/Localization/SoMall");
-            });
-
-            Configure<MultiTenancyOptions>(options =>
-            {
-                options.IsEnabled = SoMallConsts.IsMultiTenancyEnabled;
+                options.IsEnabled = MultiTenancyConsts.IsEnabled;
             });
         }
     }
