@@ -7,10 +7,23 @@ using Volo.Abp.Users.EntityFrameworkCore;
 
 namespace TT.SoMall.EntityFrameworkCore
 {
+    /* This is your actual DbContext used on runtime.
+     * It includes only your entities.
+     * It does not include entities of the used modules, because each module has already
+     * its own DbContext class. If you want to share some database tables with the used modules,
+     * just create a structure like done for AppUser.
+     *
+     * Don't use this DbContext for database migrations since it does not contain tables of the
+     * used modules (as explained above). See SoMallMigrationsDbContext for migrations.
+     */
     [ConnectionStringName("Default")]
     public class SoMallDbContext : AbpDbContext<SoMallDbContext>
     {
         public DbSet<AppUser> Users { get; set; }
+
+        /* Add DbSet properties for your Aggregate Roots / Entities here.
+         * Also map them inside SoMallDbContextModelCreatingExtensions.ConfigureSoMall
+         */
 
         public SoMallDbContext(DbContextOptions<SoMallDbContext> options)
             : base(options)
@@ -27,10 +40,7 @@ namespace TT.SoMall.EntityFrameworkCore
             builder.Entity<AppUser>(b =>
             {
                 b.ToTable("AbpUsers"); //Sharing the same table "AbpUsers" with the IdentityUser
-
-                b.ConfigureFullAudited();
-                b.ConfigureExtraProperties();
-                b.ConfigureConcurrencyStamp();
+                b.ConfigureByConvention();
                 b.ConfigureAbpUser();
 
                 //Moved customization to a method so we can share it with the SoMallMigrationsDbContext class
