@@ -1,9 +1,11 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Volo.Abp.AuditLogging.EntityFrameworkCore;
 using Volo.Abp.BackgroundJobs.EntityFrameworkCore;
+using Volo.Abp.EntityFrameworkCore;
 using Volo.Abp.EntityFrameworkCore.SqlServer;
 using Volo.Abp.FeatureManagement.EntityFrameworkCore;
 using Volo.Abp.Identity.EntityFrameworkCore;
+using Volo.Abp.IdentityServer.EntityFrameworkCore;
 using Volo.Abp.Modularity;
 using Volo.Abp.PermissionManagement.EntityFrameworkCore;
 using Volo.Abp.SettingManagement.EntityFrameworkCore;
@@ -14,10 +16,11 @@ namespace TT.SoMall.EntityFrameworkCore
     [DependsOn(
         typeof(SoMallDomainModule),
         typeof(AbpIdentityEntityFrameworkCoreModule),
+        typeof(AbpIdentityServerEntityFrameworkCoreModule),
         typeof(AbpPermissionManagementEntityFrameworkCoreModule),
         typeof(AbpSettingManagementEntityFrameworkCoreModule),
         typeof(AbpEntityFrameworkCoreSqlServerModule),
-        typeof(BackgroundJobsEntityFrameworkCoreModule),
+        typeof(AbpBackgroundJobsEntityFrameworkCoreModule),
         typeof(AbpAuditLoggingEntityFrameworkCoreModule),
         typeof(AbpTenantManagementEntityFrameworkCoreModule),
         typeof(AbpFeatureManagementEntityFrameworkCoreModule)
@@ -28,8 +31,16 @@ namespace TT.SoMall.EntityFrameworkCore
         {
             context.Services.AddAbpDbContext<SoMallDbContext>(options =>
             {
-                //Remove "includeAllEntities: true" to create default repositories only for aggregate roots
+                /* Remove "includeAllEntities: true" to create
+                 * default repositories only for aggregate roots */
                 options.AddDefaultRepositories(includeAllEntities: true);
+            });
+
+            Configure<AbpDbContextOptions>(options =>
+            {
+                /* The main point to change your DBMS.
+                 * See also SoMallMigrationsDbContextFactory for EF Core tooling. */
+                options.UseSqlServer();
             });
         }
     }
