@@ -1,9 +1,15 @@
 ﻿@page
-@using TT.SoMall.Pages
-@inherits SoMallPageBase
-@model IndexModel
+@inherits TT.SoMall.Web.Pages.SoMallPage
+@model TT.SoMall.Web.Pages.IndexModel
 @section styles {
-    <link href="~/pages/index.css" rel="stylesheet" />
+    <abp-style-bundle>
+        <abp-style src="/Pages/Index.css" />
+    </abp-style-bundle>
+}
+@section scripts {
+    <abp-script-bundle>
+        <abp-script src="/Pages/Index.js" />
+    </abp-script-bundle>
 }
 <abp-card>
     <abp-card-header>@L["Welcome"]</abp-card-header>
@@ -11,9 +17,11 @@
         <p>@L["LongWelcomeMessage"]</p>
         @if (!CurrentUser.IsAuthenticated)
         {
-            <p><a abp-button="Primary" href="/Account/Login"><i class="fa fa-sign-in"></i> @L["Login"]</a></p>
+            <form method="POST">
+                <input type="submit" asp-page-handler="Login" value="LOGIN" class="btn btn-login" />
+            </form>
         }
-        <hr/>
-        <p class="text-right"><a href="http://abp.io?ref=tmpl" target="_blank">abp.io</a></p>
+        <hr />
+        <p class="text-right"><a href="https://abp.io?ref=tmpl" target="_blank">abp.io</a></p>
     </abp-card-body>
 </abp-card>
