@@ -1,10 +1,18 @@
-﻿namespace TT.SoMall.Pages
+﻿using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authentication;
+
+namespace TT.SoMall.Web.Pages
 {
-    public class IndexModel : SoMallPageModelBase
+    public class IndexModel : SoMallPageModel
     {
         public void OnGet()
         {
             
+        }
+
+        public async Task OnPostLoginAsync()
+        {
+            await HttpContext.ChallengeAsync("oidc");
         }
     }
 }
