@@ -5,4 +5,4 @@
     mappings: {
         
     }
-}
+};
